requirements.txt
torch
torchvision
Pillow
opencv-python
PyWavelets
numpy
scikit-learn
scikit-image
matplotlib
plotly
pandas
mlflow
git+https://github.com/jacobgil/pytorch-grad-cam.git